Make third party smcpp a self contained library
PD source only now includes a very thin interface, and smcpp code can be compiled completely independently, without project warnings etc
